+++ src/App.css
@@ -3,7 +3,7 @@ h3{
 }
 
 html{
-    background-color: red;
+    background-color: rgb(0, 0, 0);
 }
 
 .App{
