+++ src/components/Layout.js
@@ -1,0 +1,12 @@
+import React from "react";
+import Game from "./Game";
+
+function Layout() {
+  return (
+    <div className="layout">
+      <Game />
+    </div>
+  );
+}
+
+export default Layout;
